fix(engine): guard audio/media sample-index conversions against Int32 overflow
Several audio paths computed a sample index with
`(int)(absoluteTime.TotalSeconds * sampleRate)`. The unchecked double->int
cast silently yields int.MinValue once the product exceeds int.MaxValue
(~12.4h @ 48kHz, ~6.2h @ 96kHz, ~3.1h @ 192kHz). On the render/export path
this fed a wrong audio offset (SourceNode) and a wrong animated-speed time
evaluation (SpeedNode + SpeedIntegrator) with no warning -- audio glitch /
desync on long timelines at high sample rates.

Add AudioMath.TimeToSampleIndex(TimeSpan, int) -> long (the existing stateless
audio-math home) and route the reachable conversions through it:

- SourceNode: long start/length; out-of-int-range returns the already-allocated
  silence buffer (Read/ReadAudio are int-bounded regardless).
- SpeedNode: long startInSamples (downstream is double division).
- SpeedIntegrator.Integrate: sample-domain math (targetInSamples /
  secStartInSamples and the remainder loop) in long; seconds + cache stay int.
- SoundSource.ToSamples: long + clamp to [0, int.MaxValue] (latent -- no
  absolute caller today, but same bug shape behind the Read(TimeSpan) overloads).

ClipNode / DelayNode / LimiterNode / AudioProcessContext.GetSampleForTime are
bounded and not reachable; left as int, with an invariant note on ClipNode.
SourceSound.Thumbnails (reachable but a separate UI path needing a Read(int)
boundary fix) is filed as a follow-up rather than silently left.

Out of the overflow regime the behavior is unchanged. New AudioMath unit tests
cover the overflow regime; existing SpeedNode/DelayNode tests stay green.

Refs: Project #9 "AI Review" item "Audio nodes: unchecked int overflow in time->sample conversion"

+++ src/Beutl.Engine/Media/SpeedIntegrator.cs
@@ -117,11 +117,14 @@ public sealed class SpeedIntegrator : IDisposable
             _integralCache![sec + 1] = sum;
         }
 
-        // Integrate the remainder from the target second to the exact target time
-        int targetInSamples = (int)(timeSpan.TotalSeconds * _sampleRate);
-        int secStartInSamples = targetSec * _sampleRate;
+        // Integrate the remainder from the target second to the exact target time. Sample-domain
+        // math uses long: an absolute global time past int.MaxValue samples (long timelines at high
+        // sample rates) would wrap an unchecked (int) cast to a negative index and silently return
+        // a wrong (too-small) source start. targetSec stays int — it is in seconds (~68 yr ceiling).
+        long targetInSamples = (long)(timeSpan.TotalSeconds * _sampleRate);
+        long secStartInSamples = (long)targetSec * _sampleRate;
 
-        for (int i = secStartInSamples; i < targetInSamples; i++)
+        for (long i = secStartInSamples; i < targetInSamples; i++)
         {
             double t = i / (double)_sampleRate;
             float speed = animation.Interpolate(TimeSpan.FromSeconds(t));
